Abstract away ethtool counters
Move parsing and handling of ethtool counters from stas.rs to parser.rs.
Add a new trait, CounterRule, to support the abstraction. The one salient
function in the trait is counters(), which returns a vector of counter
immediate values.

Unit parsing was moved to parser.rs as well, and is now in the competence
of the particular parser. Different parsers will have different ideas about
where they allow units. E.g. for ethtool it makes sense to sprinkle them
liberally between counter matches, but that wouldn't make sense for e.g.
qdisc.

CounterKey is still pretty much married to ethtool and needs to be
abstracted away next.