Change a counter NS to &'static str
Parsers should know what their name is, so do not need to allocate a full
string.

This also removes an xxx note about the structure being ethtool-centric.
The front end has to put the key values to different columns, so they
cannot be fully opaque blobs, there is meaning to them. If we need
something more generic in the future, it will be invented as needed. For
now this is plenty fine.